src/test/java/ru/yandex/practicum/filmorate/storage/dao/film/GenreDaoImplTest.java
package ru.yandex.practicum.filmorate.storage.dao.film;

import lombok.RequiredArgsConstructor;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.jdbc.Sql;

import static org.junit.jupiter.api.Assertions.assertEquals;

@SpringBootTest
@AutoConfigureTestDatabase
@RequiredArgsConstructor(onConstructor_ = @Autowired)
@Sql(scripts = "classpath:data.sql", executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
class GenreDaoImplTest {

    private final GenreDaoImpl genreDao;

    @Test
    void getGenres() {
        assertEquals(6, genreDao.getGenres().size());
    }

    @Test
    void getGenreById() {
        assertEquals("Комедия", genreDao.getGenreById(1).getName());
    }
}